[wit][wip] support for different environments
