+++ lib/config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  DEFAULT_API_VERSION: '20160516',
+  DEFAULT_MAX_STEPS: 5,
+  DEFAULT_WIT_URL: 'https://api.wit.ai'
+};
