+++ lib/wit.js
@@ -1,13 +1,14 @@
 'use strict';
 
-const fetch = require('node-fetch');
-const readline = require('readline');
-const uuid = require('node-uuid');
+const {
+  DEFAULT_API_VERSION,
+  DEFAULT_MAX_STEPS,
+  DEFAULT_WIT_URL
+} = require('./config');
+const fetch = require('isomorphic-fetch');
 const log = require('./log');
+const uuid = require('node-uuid');
 
-const DEFAULT_API_VERSION = '20160516';
-const DEFAULT_MAX_STEPS = 5;
-const DEFAULT_WIT_URL = 'https://api.wit.ai';
 const learnMore = 'Learn more at https://wit.ai/docs/quickstart';
 
 function Wit(opts) {
@@ -127,35 +128,6 @@ function Wit(opts) {
       continueRunActions(sessionId, message, context, steps)
     );
   };
-
-  this.interactive = (initContext, maxSteps) => {
-    if (!actions) throwMustHaveActions();
-    let context = typeof initContext === 'object' ? initContext : {};
-    const sessionId = uuid.v1();
-    const steps = maxSteps ? maxSteps : DEFAULT_MAX_STEPS;
-    const rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout,
-    });
-    rl.setPrompt('> ');
-    const prompt = () => {
-      rl.prompt();
-      rl.write(null, {ctrl: true, name: 'e'});
-    };
-    prompt();
-    rl.on('line', (line) => {
-      line = line.trim();
-      if (!line) {
-        return prompt();
-      }
-      this.runActions(sessionId, line, context, steps)
-      .then((ctx) => {
-        context = ctx;
-        prompt();
-      })
-      .catch(logger.error)
-    });
-  };
 };
 
 const makeWitResponseHandler = (logger, endpoint) => {
@@ -263,6 +235,4 @@ const clone = (obj) => {
   }
 };
 
-module.exports = {
-  Wit,
-};
+module.exports = Wit;
